Added test files for XHTML structure import. Added ImportStructure class, menu item, user prompts
There is a problem with the import and TOC sync: the TOC remembers node data as "Untitled Section" (the name before one gets assigned), and the project itself is using the text label (e.g. "first section").

Try with application\misc\test xhtml import\importTest-valid-simple-2.html to see this behavior.

Another problem is getting the inner text of an element that has children (e.g. <h1><a href="">the text</a></h1>, but I suspect this is more easily solved.